refactor of decision engine
