+++ CMakeLists.txt
@@ -10,6 +10,7 @@ project(sample_project CXX C)
 # Core library. *.cpp should be added here.
 add_library(core
   ./src/core.cpp
+  ./src/decision_engine.cpp
   ./src/action.cpp)
 
 # Main entry point.
