+++ src/main.cpp
@@ -4,10 +4,10 @@
 #include <vector>
 
 #include "core.h"
+#include "decision_engine.h"
 
 int main(int argc, char** argv) {
   printf("creating CVC\n");
-  std::unique_ptr<DecisionEngine> d;
   std::vector<std::unique_ptr<Character>> c;
   c.push_back(std::make_unique<Character>(1, 100.0));
   c.push_back(std::make_unique<Character>(2, 100.0));
@@ -17,10 +17,11 @@ int main(int argc, char** argv) {
   std::mt19937 random_generator(rd());
 
   FILE* action_log = fopen("/tmp/action_log", "a");
-  CVC cvc(std::move(d), std::move(c), random_generator, action_log);
+  CVC cvc(std::move(c), random_generator);
+  DecisionEngine d(&cvc, action_log);
 
   printf("running the game loop\n");
-  cvc.GameLoop();
+  d.GameLoop();
 
   fclose(action_log);
